Clean up remaining byte casts from 39081e62c95450a098004cef335acc54ab32a795

+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/extent/ExtentHeightCacher.java
@@ -46,7 +46,7 @@ public class ExtentHeightCacher extends PassthroughExtent {
         }
         int result = cacheHeights[index] & 0xFF;
         if (result == minY) {
-            cacheHeights[index] = (byte) (result = lastY = super
+            cacheHeights[index] = (short) (result = lastY = super
                     .getNearestSurfaceTerrainBlock(x, z, lastY, minY, maxY));
         }
         return result;
